requirements.txt
pandas>=2.2
numpy==1.26.4 
scikit-learn>=1.5
lightgbm>=4.3
pyyaml>=6.0
joblib==1.4.2
holidays>=0.41
pyarrow==16.1.0